split DateRange into DateRangeForm and DateRangeDisplay
