+++ widget-src/components/log/LogDisplay.tsx
@@ -1,7 +1,7 @@
 import { ChangeLog, ChangeLogState } from '../../types/ChangeLog';
 import { COLOR, FONT, GAP, PADDING } from '../../utilities/Styles';
 import { Type } from './Type';
-import { DateRange } from './DateRange';
+import { DateRange } from './DateRangeDisplay';
 import { Button } from '../Button';
 import { ActionEditIcon } from '../../svgs/ActionEditIcon';
 import { LinkList } from './LinkList';
@@ -67,7 +67,6 @@ export const ChangeLogDisplay = ({
         </Text>
 
         <DateRange
-          changeLog={changeLog}
           timestamp={changeLog.createdDate}
           editedTimestamp={changeLog.editedDate}
           editCount={changeLog.editCount}
